# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XLS-66 Lending Protocol Reference App
  description: |
    HTTP API for the XRPL Lending Protocol (XLS-66) + Single Asset Vault
    (XLS-65) reference implementation. Every route that mutates on-chain
    state maps 1:1 to an XRPL transaction and re-queries the ledger before
    returning.

    ## Conventions

    - **Auth** — handled by Auth0 Universal Login. Clients obtain a session
      by visiting `/auth/login` (browser only) — there is no programmatic
      login endpoint. The session is stored in an httpOnly cookie `xls66-auth`.
      Every `/api/*` route is gated by `src/middleware.ts` and returns
      `401 Unauthorized` without a valid cookie.
    - **Amounts** — drops (integer string) for XRP; decimal string for IOU
      and MPT. See the README for the scaling conventions.
    - **Rates** — UI-facing rates are % or basis points; on-ledger they are
      1/10 bps. The API accepts the UI form and scales at the boundary.
    - **Content type** — `application/json` for all request/response bodies.
  version: 1.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: XLS-66 reference app
servers:
  - url: http://localhost:3000
    description: Local dev server
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xls66-auth
      description: httpOnly cookie set by `POST /api/session`. Value is the Session document's MongoDB ObjectId.
...
  schemas:
...
    VaultAsset:
      type: object
      properties:
        type:
          type: string
          enum: [XRP, IOU, MPT]
        currency: { type: string }
        issuer: { type: string }
        mptIssuanceId: { type: string }
    ShareMetadata:
      description: XLS-89 compressed MPT metadata for the vault-share token.
      type: object
      properties:
        ticker: { type: string, maxLength: 6 }
        name: { type: string, maxLength: 64 }
        description: { type: string, maxLength: 256 }
        icon: { type: string, maxLength: 128 }
        issuerName: { type: string, maxLength: 64 }
    VaultCreateOptions:
      type: object
      properties:
        asset: { $ref: "#/components/schemas/VaultAsset" }
        name: { type: string, maxLength: 64 }
        website: { type: string, maxLength: 128 }
        assetsMaximum:
          type: string
          description: "Deposit cap. Drops for XRP; asset units for IOU/MPT."
        nonTransferableShares:
          type: boolean
          description: "Sets tfVaultShareNonTransferable flag."
        shareMetadata: { $ref: "#/components/schemas/ShareMetadata" }
    Vault:
      description: MongoDB cache row for a vault. Combined with `onLedger` for a full view.
      type: object
      required: [_id, sessionId, vaultId, ownerAddress]
      properties:
        _id: { type: string }
        sessionId: { type: string }
        vaultId: { type: string }
        ownerAddress: { type: string }
        asset:
          type: object
          properties:
            currency: { type: string }
            issuer: { type: string }
            mptIssuanceId: { type: string }
        totalDeposited: { type: string }
        sharesMinted: { type: string }
        status: { type: string, enum: [active, deleted] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        onLedger:
          description: Live `vault_info` response, amount fields unscaled for MPT.
          type: object
          additionalProperties: true
...
    TxResult:
      description: Pass-through of the rippled `tx` submission result. Only the hash is canonical; everything else is for debugging.
      type: object
      additionalProperties: true
      properties:
        hash: { type: string }
        ledger_index: { type: integer }
        meta:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid `xls66-auth` cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Referenced entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Unhandled error or rippled failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /api/vault:
    post:
      tags: [Vault]
      summary: Create a vault (VaultCreate)
      description: |
        Submits a `VaultCreate` transaction. If `vaultOptions.asset.type` is
        `IOU` or `MPT`, the server first runs `setupIOU` / `setupMPT` — this
        creates trustlines / MPT issuance, authorizes the three role wallets,
        and distributes demo stake to all of them. The resulting asset is
        persisted on the session as `issuedToken`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vaultOptions: { $ref: "#/components/schemas/VaultCreateOptions" }
      responses:
        "201":
          description: Vault created
          content:
            application/json:
              schema:
                type: object
                properties:
                  vault: { $ref: "#/components/schemas/Vault" }
                  result: { $ref: "#/components/schemas/TxResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
    get:
      tags: [Vault]
      summary: List active vaults
      description: Returns every active vault for the session, enriched with live `vault_info`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  vaults:
                    type: array
                    items: { $ref: "#/components/schemas/Vault" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/vault/deposit:
    post:
      tags: [Vault]
      summary: VaultDeposit
      description: Supply liquidity. Provide `amountDrops` for XRP vaults, `tokenAmount` (decimal) for IOU/MPT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vaultId]
              properties:
                vaultId: { type: string }
                amountDrops: { type: string, description: "XRP drops (integer string)" }
                tokenAmount: { type: string, description: "IOU/MPT decimal" }
      responses:
        "200":
          description: Deposit submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/vault/delete:
    post:
      tags: [Vault]
      summary: Full teardown (loans → broker → vault)
      description: |
        Defaults every active loan, deletes them, withdraws first-loss
        cover, deletes the broker, then deletes the vault. Any step that
        fails individually is logged but doesn't block later steps — the
        final `VaultDelete` surfaces the real error if state is still dirty.
      responses:
        "200":
          description: Vault deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "400":
          description: "Vault not empty (assets still deposited)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/vault/history:
    get:
      tags: [Vault]
      summary: Deposit / withdraw audit log + PNL summary
      parameters:
        - name: vaultId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        _id: { type: string }
                        type: { type: string, enum: [deposit, withdraw] }
                        amountDrops: { type: string, description: "Unscaled for MPT on read." }
                        txHash: { type: string }
                        createdAt: { type: string, format: date-time }
                  summary:
                    type: object
                    properties:
                      totalDeposited: { type: string }
                      totalWithdrawn: { type: string }
                      netInvested: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...